Moved JsonSupport to new webserver module. Fixed a few compile errors. Fixed compile issues caused by webserver being split to a separate project. WebServer now starts and stops correctly as a separate module.
