+++ core/src/main/kotlin/net/corda/core/messaging/CordaRPCOps.kt
@@ -118,7 +118,7 @@ interface CordaRPCOps : RPCOps {
      */
     fun uploadAttachment(jar: InputStream): SecureHash
 
-    @Suppress("DEPRECATION")
+    // TODO: Remove this from the interface
     @Deprecated("This service will be removed in a future milestone")
     fun uploadFile(dataType: String, name: String?, file: InputStream): String
