+++ docs/source/example-code/build.gradle
@@ -49,6 +49,7 @@ dependencies {
     }
 
     runtime project(path: ":node:capsule", configuration: 'runtimeArtifacts')
+    runtime project(path: ":node:webserver", configuration: 'runtimeArtifacts')
 }
 
 mainClassName = "net.corda.docs.ClientRpcTutorialKt"
