+++ gradle-plugins/publish-utils/src/main/groovy/net/corda/plugins/ProjectPublishExtension.groovy
@@ -9,4 +9,8 @@ class ProjectPublishExtension {
      * True when we do not want to publish default Java components
      */
     Boolean disableDefaultJar = false
+    /**
+     * True if publishing a WAR instead of a JAR. Forces disableDefaultJAR to "true" when true
+     */
+    Boolean publishWar = false
 }
