+++ gradle-plugins/publish-utils/src/main/groovy/net/corda/plugins/PublishTasks.groovy
@@ -60,8 +60,10 @@ class PublishTasks implements Plugin<Project> {
     void configureMavenPublish(BintrayConfigExtension bintrayConfig) {
         project.apply([plugin: 'maven-publish'])
         project.publishing.publications.create(publishName, MavenPublication) {
-            if(!publishConfig.disableDefaultJar) {
+            if(!publishConfig.disableDefaultJar && !publishConfig.publishWar) {
                 from project.components.java
+            } else if(publishConfig.publishWar) {
+                from project.components.web
             }
             groupId project.group
             artifactId publishName
