+++ node/build.gradle
@@ -88,30 +88,12 @@ dependencies {
         exclude group: "asm"
     }
 
-    // Force commons logging to version 1.2 to override Artemis, which pulls in 1.1.3 (ARTEMIS-424)
-    compile "commons-logging:commons-logging:1.2"
-
-    // Web stuff: for HTTP[S] servlets
-    compile "org.eclipse.jetty:jetty-servlet:${jetty_version}"
-    compile "org.eclipse.jetty:jetty-webapp:${jetty_version}"
-    compile "javax.servlet:javax.servlet-api:3.1.0"
-    compile "org.jolokia:jolokia-agent-war:$jolokia_version"
+    // For adding serialisation of file upload streams to RPC
+    // TODO: Remove this dependency and the code that requires it
     compile "commons-fileupload:commons-fileupload:1.3.2"
 
-    // Jersey for JAX-RS implementation for use in Jetty
-    compile "org.glassfish.jersey.core:jersey-server:${jersey_version}"
-    compile "org.glassfish.jersey.containers:jersey-container-servlet-core:${jersey_version}"
-    compile "org.glassfish.jersey.containers:jersey-container-jetty-http:${jersey_version}"
-    // NOTE there is a Jackson version clash between jersey-media-json-jackson (v2.5.4) and jackson-module-kotlin (v.2.5.5)
-    // Have not found an Issue in the issue tracker for Jersey for this issue
-    compile ("org.glassfish.jersey.media:jersey-media-json-jackson:${jersey_version}") {
-        exclude group: 'com.fasterxml.jackson.core', module: 'jackson-annotations'
-        exclude group: 'com.fasterxml.jackson.core', module: 'jackson-databind'
-        exclude group: 'com.fasterxml.jackson.core', module: 'jackson-core'
-    }
-    compile ("com.fasterxml.jackson.module:jackson-module-kotlin:${jackson_version}") {
-        exclude group: 'com.fasterxml.jackson.core', module: 'jackson-annotations'
-    }
+    // Force commons logging to version 1.2 to override Artemis, which pulls in 1.1.3 (ARTEMIS-424)
+    compile "commons-logging:commons-logging:1.2"
     compile "com.fasterxml.jackson.core:jackson-annotations:${jackson_version}"
 
     // Coda Hale's Metrics: for monitoring of key statistics
