+++ node/src/main/kotlin/net/corda/node/ArgsParser.kt
@@ -28,7 +28,6 @@ class ArgsParser {
             .withValuesConvertedBy(object : EnumConverter<Level>(Level::class.java) {})
             .defaultsTo(Level.INFO)
     private val logToConsoleArg = optionParser.accepts("log-to-console", "If set, prints logging to the console as well as to a file.")
-    private val isWebserverArg = optionParser.accepts("webserver")
     private val isRegistrationArg = optionParser.accepts("initial-registration", "Start initial node registration with Corda network to obtain certificate from the permissioning server.")
     private val helpArg = optionParser.accepts("help").forHelp()
 
@@ -42,9 +41,8 @@ class ArgsParser {
         val help = optionSet.has(helpArg)
         val loggingLevel = optionSet.valueOf(loggerLevel)
         val logToConsole = optionSet.has(logToConsoleArg)
-        val isWebserver = optionSet.has(isWebserverArg)
         val isRegistration = optionSet.has(isRegistrationArg)
-        return CmdLineOptions(baseDirectory, configFile, help, loggingLevel, logToConsole, isWebserver, isRegistration)
+        return CmdLineOptions(baseDirectory, configFile, help, loggingLevel, logToConsole, isRegistration)
     }
 
     fun printHelp(sink: PrintStream) = optionParser.printHelpOn(sink)
@@ -55,7 +53,6 @@ data class CmdLineOptions(val baseDirectory: Path,
                           val help: Boolean,
                           val loggingLevel: Level,
                           val logToConsole: Boolean,
-                          val isWebserver: Boolean,
                           val isRegistration: Boolean) {
     fun loadConfig(allowMissingConfig: Boolean = false, configOverrides: Map<String, Any?> = emptyMap()): Config {
         return ConfigHelper.loadConfig(baseDirectory, configFile, allowMissingConfig, configOverrides)
