+++ node/src/main/kotlin/net/corda/node/Corda.kt
@@ -10,7 +10,6 @@ import net.corda.node.services.config.FullNodeConfiguration
 import net.corda.node.utilities.ANSIProgressObserver
 import net.corda.node.utilities.registration.HTTPNetworkRegistrationService
 import net.corda.node.utilities.registration.NetworkRegistrationHelper
-import net.corda.node.webserver.WebServer
 import org.fusesource.jansi.Ansi
 import org.fusesource.jansi.AnsiConsole
 import org.slf4j.LoggerFactory
@@ -61,8 +60,7 @@ fun main(args: Array<String>) {
 
     drawBanner()
 
-    val logDir = if (cmdlineOptions.isWebserver) "logs/web" else "logs"
-    System.setProperty("log-path", (cmdlineOptions.baseDirectory / logDir).toString())
+    System.setProperty("log-path", (cmdlineOptions.baseDirectory / "logs").toString())
 
     val log = LoggerFactory.getLogger("Main")
     printBasicNodeInfo("Logs can be found in", System.getProperty("log-path"))
@@ -94,39 +92,26 @@ fun main(args: Array<String>) {
     log.info("VM ${info.vmName} ${info.vmVendor} ${info.vmVersion}")
     log.info("Machine: ${InetAddress.getLocalHost().hostName}")
     log.info("Working Directory: ${cmdlineOptions.baseDirectory}")
-    if (cmdlineOptions.isWebserver) {
-        log.info("Starting as webserver on ${conf.webAddress}")
-    } else {
-        log.info("Starting as node on ${conf.artemisAddress}")
-    }
+    log.info("Starting as node on ${conf.artemisAddress}")
 
     try {
         cmdlineOptions.baseDirectory.createDirectories()
 
-        // TODO: Webserver should be split and start from inside a WAR container
-        if (!cmdlineOptions.isWebserver) {
-            val node = conf.createNode()
-            node.start()
-            printPluginsAndServices(node)
+        val node = conf.createNode()
+        node.start()
+        printPluginsAndServices(node)
 
-            node.networkMapRegistrationFuture.success {
-                val elapsed = (System.currentTimeMillis() - startTime) / 10 / 100.0
-                printBasicNodeInfo("Node started up and registered in $elapsed sec")
-
-                if (renderBasicInfoToConsole)
-                    ANSIProgressObserver(node.smm)
-            } failure {
-                log.error("Error during network map registration", it)
-                exitProcess(1)
-            }
-            node.run()
-        } else {
-            val server = WebServer(conf)
-            server.start()
+        node.networkMapRegistrationFuture.success {
             val elapsed = (System.currentTimeMillis() - startTime) / 10 / 100.0
-            printBasicNodeInfo("Webserver started up in $elapsed sec")
-            server.run()
+            printBasicNodeInfo("Node started up and registered in $elapsed sec")
+
+            if (renderBasicInfoToConsole)
+                ANSIProgressObserver(node.smm)
+        } failure {
+            log.error("Error during network map registration", it)
+            exitProcess(1)
         }
+        node.run()
     } catch (e: Exception) {
         log.error("Exception during node startup", e)
         exitProcess(1)
