+++ node/src/main/kotlin/net/corda/node/internal/Node.kt
@@ -33,7 +33,6 @@ import java.nio.file.Files
 import java.nio.file.Paths
 import java.time.Clock
 import javax.management.ObjectName
-import javax.servlet.*
 import kotlin.concurrent.thread
 
 /**
@@ -287,18 +286,6 @@ class Node(override val configuration: FullNodeConfiguration,
         val ourProcessID: String = ManagementFactory.getRuntimeMXBean().name.split("@")[0]
         f.setLength(0)
         f.write(ourProcessID.toByteArray())
-    }
-
-    // Servlet filter to wrap API requests with a database transaction.
-    private class DatabaseTransactionFilter(val database: Database) : Filter {
-        override fun doFilter(request: ServletRequest, response: ServletResponse, chain: FilterChain) {
-            databaseTransaction(database) {
-                chain.doFilter(request, response)
-            }
-        }
-
-        override fun init(filterConfig: FilterConfig?) {}
-        override fun destroy() {}
     }
 }
