+++ publish.properties
@@ -1,1 +1,1 @@
-gradlePluginsVersion=0.8.2
+gradlePluginsVersion=0.8.3
