+++ samples/irs-demo/build.gradle
@@ -49,8 +49,10 @@ dependencies {
 
     // Corda integration dependencies
     runtime project(path: ":node:capsule", configuration: 'runtimeArtifacts')
+    runtime project(path: ":node:webserver", configuration: 'runtimeArtifacts')
     compile project(':core')
     compile project(':finance')
+    compile project(':node:webserver')
     compile project(':test-utils')
 
     // Javax is required for webapis
