+++ samples/network-visualiser/build.gradle
@@ -22,6 +22,7 @@ dependencies {
 
     // Corda integration dependencies
     runtime project(path: ":node:capsule", configuration: 'runtimeArtifacts')
+    runtime project(path: ":node:webserver", configuration: 'runtimeArtifacts')
     compile project(':core')
     compile project(':finance')
     testCompile project(':test-utils')
