+++ samples/simm-valuation-demo/build.gradle
@@ -54,6 +54,7 @@ dependencies {
 
     // Corda integration dependencies
     runtime project(path: ":node:capsule", configuration: 'runtimeArtifacts')
+    runtime project(path: ":node:webserver", configuration: 'runtimeArtifacts')
     compile project(':core')
     compile project(':node')
     compile project(':finance')
