+++ samples/trader-demo/build.gradle
@@ -46,6 +46,7 @@ dependencies {
 
     // Corda integration dependencies
     runtime project(path: ":node:capsule", configuration: 'runtimeArtifacts')
+    runtime project(path: ":node:webserver", configuration: 'runtimeArtifacts')
     compile project(':core')
     compile project(':finance')
     compile project(':test-utils')
